顧客検索画面追加実装

+++ my-spring-boot-app/src/main/java/com/example/myapp/entity/Organization.java
@@ -1,16 +1,16 @@
 package com.example.myapp.entity;
 
 public class Organization {
-    private Long organizationId;
+    private Integer organizationId;
     private String organizationName;
     private Integer upperOrganization;
 
     // Getters and Setters
-    public Long getOrganizationId() {
+    public Integer getOrganizationId() {
         return organizationId;
     }
 
-    public void setOrganizationId(Long organizationId) {
+    public void setOrganizationId(Integer organizationId) {
         this.organizationId = organizationId;
     }
